Update flagColors.ino
Added more meaningful event names and data that publish based on PIR and SoundDetector output. Also created a function "publishEvent" from previous function "publishMyEvent" that takes a string eventID and a string data. The eventID is added to the eventName variable to help specify which events each particle (or busStop) should subscribe to and act upon. The data ensures that there is a parsable String regardless of what event is being subscribed to.